+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 - Added an option to change the highlight color in the research screen
 
 ### Changed
+- You can now resize the sidebar in the research screen
 - Somewhat improved the config layout
 - Pinned researches that can't be researched are now automatically unpinned
 
 ### Fixed
+- Fixed not being able to drag to scroll in the research tree view
 - Fixed a minor data component predicate issue
 - Fixed research reward icons being on top of each other in some cases